Create no prefix set - hacker rank
There is a given list of strings where each string contains only lowercase letters from a-j , inclusive. The set of strings is said to be a GOOD SET if no string is a prefix of another string. In this case, print GOOD SET. Otherwise, print BAD SET on the first line followed by the string being checked.